refactor existing table component to create common component
