+++ src/app/static/src/app/components/modelOutput/ModelOutput.vue
@@ -28,7 +28,7 @@
                     @update-colour-scales="updateOutputColourScales({payload: $event})"></choropleth>
                 <div class="row mt-2">
                     <div class="col-md-3"></div>
-                    <table-view class="col-md-9"
+                    <area-indicators-table class="col-md-9"
                                 :tabledata="chartdata"
                                 :area-filter-id="areaFilterId"
                                 :filters="choroplethFilters"
@@ -37,7 +37,7 @@
                                 :selections="choroplethSelections"
 
                                 :selectedFilterOptions="choroplethSelections.selectedFilterOptions"
-                    ></table-view>
+                    ></area-indicators-table>
                 </div>
             </div>
 
@@ -51,7 +51,7 @@
                     @update="updateBarchartSelections({payload: $event})"></bar-chart-with-filters>
                 <div class="row mt-2">
                     <div class="col-md-3"></div>
-                    <table-view class="col-md-9"
+                    <area-indicators-table class="col-md-9"
                                 :tabledata="chartdata"
                                 :area-filter-id="areaFilterId"
                                 :filters="barchartFilters"
@@ -60,7 +60,7 @@
                                 :selections="barchartSelections"
 
                                 :selectedFilterOptions="barchartSelections.selectedFilterOptions"
-                    ></table-view>
+                    ></area-indicators-table>
                 </div>
             </div>
 
@@ -76,7 +76,7 @@
                              @update-size-scales="updateOutputBubbleSizeScales({payload: $event})"></bubble-plot>
                 <div class="row mt-2">
                     <div class="col-md-3"></div>
-                    <table-view class="col-md-9"
+                    <area-indicators-table class="col-md-9"
                                 :tabledata="chartdata"
                                 :area-filter-id="areaFilterId"
                                 :filters="bubblePlotFilters"
@@ -85,7 +85,7 @@
                                 :selections="bubblePlotSelections"
 
                                 :selectedFilterOptions="bubblePlotSelections.selectedFilterOptions"
-                    ></table-view>
+                    ></area-indicators-table>
                 </div>
             </div>
         </div>
@@ -97,7 +97,7 @@
     import Vue from "vue";
     import Choropleth from "../plots/choropleth/Choropleth.vue";
     import BubblePlot from "../plots/bubble/BubblePlot.vue";
-    import TableView from "../plots/table/Table.vue";
+    import AreaIndicatorsTable from "../plots/table/AreaIndicatorsTable.vue";
     import {BarchartIndicator, Filter, FilterConfig, FilterOption} from "@reside-ic/vue-charts/src/bar/types";
     import {BarChartWithFilters} from "@reside-ic/vue-charts";
 
@@ -240,7 +240,7 @@
             BarChartWithFilters,
             BubblePlot,
             Choropleth,
-            TableView
+            AreaIndicatorsTable
         }
     })
 </script>
