+++ src/app/static/src/app/components/surveyAndProgram/SurveyAndProgram.vue
@@ -46,14 +46,14 @@
                             @update="updateChoroplethSelections({payload: $event})"
                             @update-colour-scales="updateSAPColourScales({payload: [selectedDataType, $event]})"></choropleth>
                 <div>
-                    <table-view :tabledata="data"
+                    <area-indicators-table :tabledata="data"
                                 :area-filter-id="areaFilterId"
                                 :filters="filters"
                                 :countryAreaFilterOption="countryAreaFilterOption"
                                 :indicators="filterTableIndicators"
                                 :selections="plottingSelections"
                                 :selectedFilterOptions="plottingSelections.selectedFilterOptions"
-                    ></table-view>
+                    ></area-indicators-table>
                 </div>
             </div>
         </div>
@@ -72,7 +72,7 @@
     import i18next from "i18next";
     import {mapGetters, mapMutations, mapState} from "vuex";
     import Choropleth from "../plots/choropleth/Choropleth.vue";
-    import TableView from "../plots/table/Table.vue";
+    import AreaIndicatorsTable from "../plots/table/AreaIndicatorsTable.vue";
     import Filters from "../plots/Filters.vue";
     import {
         Filter,
@@ -208,7 +208,7 @@
         components: {
             Choropleth,
             Filters,
-            TableView,
+            AreaIndicatorsTable,
             TreeSelect,
             GenericChart
         }
